requirements.txt
streamlit==1.29.0
tensorflow==2.12.0
numpy==1.24.3
Pillow==10.0.0